% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat.coef.EpC}
\alias{mat.coef.EpC}
\title{coefficient matrix for PCs(principal components) of Epithelium/Fibroblast}
\format{
\subsection{\code{mat.coef.EpC}}{

A data frame with 2,214(EpC)/ 2,935(Fib) rows(trajectory features merged/filtered from stage-DEGs) and 50 columns(top50 PCs)
}
}
\source{
\url{https://github.com/Ruismart/PLHT}
}
\usage{
mat.coef.EpC
}
\description{
built from mouse intestinal scRNAseq stage-WT data(P00-P07-P14-P21-P28, Epcam+ Epithelial or Pdgfra+ Mesenchymal),
used for custom PCA construction (further fit the linear regression model to map onto the time coordinate)
}
\keyword{datasets}
