% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plht.r
\name{plht}
\alias{plht}
\title{Using PLHT to measure the postnatal development level -- plht(input matrix)}
\usage{
plht(
  mat,
  design,
  stsp,
  stsp.type = "median",
  scale = TRUE,
  mat.coef.EpC = mat.coef.EpC,
  mat.coef.Fib = mat.coef.Fib,
  features.overlap = TRUE
)
}
\arguments{
\item{mat}{a normalized expression matrix:
for bulk RNAseq, should be log2(TPM/CPM/TMM+1);
for singlecell RNAseq, should be assays-'RNA/SCT'-data.}

\item{design}{A design table of meta.data containing several columns as:
sample - condition - replicate - tissue - age - batch - ...
(first three are necessary)
(design$sample must match colnames(mat))}

\item{stsp}{To specify standard sample condition in design$condition,
the final trajectory.score would set 0 for the median(/mean) of this condition.}

\item{stsp.type}{set "median"(default) or "mean" value for stsp}

\item{scale}{Whether to do zscore-transformation for the input data matrix.
For singlecell, recommend to use @scale.data instead, then set scale=FALSE;
or use another function:plht.seur which is modified for seurat object as input.}

\item{mat.coef.EpC}{coefficient matrix for PCs(principal components) of Epithelium, built-in data}

\item{mat.coef.Fib}{coefficient matrix for PCs(principal components) of Fibroblast, built-in data}

\item{features.overlap}{TRUE/FALSE: TRUE(default) to include overlapped features of EpC/Fib in the calculation.}
}
\value{
A dataframe: sample - Tc.EpC - Tc.Fib  - condition - replicate - (other design columns);
print information for the input data: EpC/Fib features   xxxx, expressed   xxxx
}
\description{
Please check https://github.com/Ruismart/PLHT for details of the method and code example to plot the result;
}
\examples{
#

}
