% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plht.seur.r
\name{plht.seur}
\alias{plht.seur}
\title{Using PLHT to measure the postnatal development level -- plht.seur(input seurat.object)}
\usage{
plht.seur(
  seur.obj,
  assay = "RNA",
  condition = "condition",
  stsp = "CTL|WT|D0|P0",
  stsp.type = "median",
  mat.coef.EpC = mat.coef.EpC,
  mat.coef.Fib = mat.coef.Fib,
  features.overlap = TRUE
)
}
\arguments{
\item{seur.obj}{a well-normalized seurat object containing at least two conditions
which are associated with postnatal development}

\item{assay}{To select assays-'RNA/SCT'-scale.data for calculation, default is "RNA"}

\item{condition}{the condition column in seur.obj@meta.data}

\item{stsp}{To specify standard sample condition in seur.obj@meta.data$condition,
the final trajectory.score would set 0 for the median(/mean) of this condition.}

\item{stsp.type}{set "median"(default) or "mean" value for stsp}

\item{mat.coef.EpC}{coefficient matrix for PCs(principal components) of Epithelium, built-in data}

\item{mat.coef.Fib}{coefficient matrix for PCs(principal components) of Fibroblast, built-in data}

\item{features.overlap}{TRUE/FALSE: TRUE(default) to include overlapped features of EpC/Fib in the calculation.}
}
\value{
same seurat object with Tc.EpC/Tc.Fib added as new columns in meta.data
}
\description{
Please check https://github.com/Ruismart/PLHT for details of the method and code example to plot the result;
}
\examples{
#

}
